can now update name
