+++ server/routes/api/users.js
@@ -24,7 +24,14 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:id', async (req, res, next) => {
   try {
-    const user = await User.findByPk(+req.params.id);
+    const user = await User.findOne({
+      where: {id: +req.params.id},
+      include: [
+        {
+          model: Address,
+        },
+      ],
+    });
     res.status(200).json(user);
   } catch (err) {
     next(err);
