+++ src/redux-store/reducers/auth-reducer.js
@@ -7,13 +7,13 @@ import {
 
 export default (state = {}, action) => {
   switch (action.type) {
+    case UPDATE_ONE_USERDETAILS:
+      return {...state, userDetails: action.userDetails};
     case GET_USER_DETAILS:
     case LOGIN_USER:
       return {...state, userDetails: action.userDetails};
     case LOGOUT_USER:
       return {...action.userDetails};
-    case UPDATE_ONE_USERDETAILS:
-      return {...state, userDetails: action.userDetails};
     default:
       return state;
   }
